feat: proxy pass

+++ helper/config.go
@@ -25,7 +25,7 @@ type config struct {
 	NonGetMode       byte          // How to deal with non-GET requests: pass|block|cache|queue
 	QueueLength      int           `json:"queue_length"` // Queue at most this number of requests for `non_get_mode=queue`. Otherwise has no effect
 	QueueSize        int           `json:"queue_size"`   // Max queue size in bytes for `non_get_mode=queue`, default 1 MB. Otherwise has no effect
-	DequeueRate      int           `json:"dequeue_rate"` // Dequeue and forward this number of queued requests per second when `non_get_mode=queue`
+	DequeueRate      float32       `json:"dequeue_rate"` // Dequeue and forward this number of queued requests per second when `non_get_mode=queue`
 	LruTime          time.Duration // track access count within this time period (minutes) for each cache entry
 	ProtectionExpire time.Duration // Fresh requests will go stale and fall into LRU list after this much of time (minutes)
 	// TODO: support cache TTL, manual cache deleting
